+++ compiler-plugin/src/main/java/io/ballerina/stdlib/ftp/plugin/FtpFunctionValidator.java
@@ -41,7 +41,11 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-import static io.ballerina.stdlib.ftp.plugin.PluginConstants.CompilationErrors.INVALID_PARAMETER;
+import static io.ballerina.compiler.syntax.tree.SyntaxKind.INTERSECTION_TYPE_DESC;
+import static io.ballerina.compiler.syntax.tree.SyntaxKind.QUALIFIED_NAME_REFERENCE;
+import static io.ballerina.stdlib.ftp.plugin.PluginConstants.CompilationErrors.INVALID_CALLER_PARAMETER;
+import static io.ballerina.stdlib.ftp.plugin.PluginConstants.CompilationErrors.INVALID_PARAMETERS;
+import static io.ballerina.stdlib.ftp.plugin.PluginConstants.CompilationErrors.INVALID_WATCHEVENT_PARAMETER;
 import static io.ballerina.stdlib.ftp.plugin.PluginConstants.CompilationErrors.METHOD_MUST_BE_REMOTE;
 import static io.ballerina.stdlib.ftp.plugin.PluginConstants.CompilationErrors.MUST_HAVE_WATCHEVENT;
 import static io.ballerina.stdlib.ftp.plugin.PluginConstants.CompilationErrors.NO_ON_FILE_CHANGE;
@@ -75,35 +79,154 @@ public class FtpFunctionValidator {
                         DiagnosticSeverity.ERROR, onFileChange.location()));
             }
             SeparatedNodeList<ParameterNode> parameters = onFileChange.functionSignature().parameters();
-            validateFunctionParameters(parameters, onFileChange);
+            validateFunctionArguments(parameters, onFileChange);
             validateReturnTypeErrorOrNil(onFileChange);
         }
     }
 
-    private void validateFunctionParameters(SeparatedNodeList<ParameterNode> parameters,
-                                            FunctionDefinitionNode functionDefinitionNode) {
+    private void validateFunctionArguments(SeparatedNodeList<ParameterNode> parameters,
+                                           FunctionDefinitionNode functionDefinitionNode) {
         if (parameters.size() == 1) {
-            ParameterNode paramNode = parameters.get(0);
-            SyntaxKind paramSyntaxKind = ((RequiredParameterNode) paramNode).typeName().kind();
-            if (paramSyntaxKind.equals(SyntaxKind.QUALIFIED_NAME_REFERENCE)) {
-                validateWatchEventParam(paramNode);
-            } else if (paramSyntaxKind.equals(SyntaxKind.INTERSECTION_TYPE_DESC)) {
-                validateIntersectionParam(paramNode);
-            } else {
-                context.reportDiagnostic(PluginUtils.getDiagnostic(
-                        INVALID_PARAMETER,
-                        DiagnosticSeverity.ERROR, paramNode.location()));
-            }
-        } else if (parameters.size() > 1) {
-            context.reportDiagnostic(PluginUtils.getDiagnostic(ONLY_PARAMS_ALLOWED,
-                    DiagnosticSeverity.ERROR, functionDefinitionNode.functionSignature().location()));
+            validateSingleArgumentScenario(parameters);
+        } else if (parameters.size() == 2) {
+            validateTwoArgumentScenario(parameters, functionDefinitionNode);
+        } else if (parameters.size() > 2) {
+            validateInvalidArgumentCountScenario(functionDefinitionNode, ONLY_PARAMS_ALLOWED);
         } else {
-            context.reportDiagnostic(PluginUtils.getDiagnostic(MUST_HAVE_WATCHEVENT,
+            validateInvalidArgumentCountScenario(functionDefinitionNode, MUST_HAVE_WATCHEVENT);
+        }
+    }
+
+    private void validateInvalidArgumentCountScenario(FunctionDefinitionNode functionDefinitionNode,
+                                                      PluginConstants.CompilationErrors compilationErrors) {
+        context.reportDiagnostic(PluginUtils.getDiagnostic(compilationErrors,
+                DiagnosticSeverity.ERROR, functionDefinitionNode.functionSignature().location()));
+    }
+
+    private void validateTwoArgumentScenario(SeparatedNodeList<ParameterNode> parameters,
+                                             FunctionDefinitionNode functionDefinitionNode) {
+        ParameterNode firstParamNode = parameters.get(0);
+        ParameterNode secondParamNode = parameters.get(1);
+        SyntaxKind firstParamSyntaxKind = ((RequiredParameterNode) firstParamNode).typeName().kind();
+        SyntaxKind secondParamSyntaxKind = ((RequiredParameterNode) secondParamNode).typeName().kind();
+
+        if (firstParamSyntaxKind.equals(QUALIFIED_NAME_REFERENCE) &&
+                secondParamSyntaxKind.equals(QUALIFIED_NAME_REFERENCE)) {
+            validateCallerAndWachEvent(firstParamNode, secondParamNode);
+        } else if (firstParamSyntaxKind.equals(INTERSECTION_TYPE_DESC) &&
+                secondParamSyntaxKind.equals(INTERSECTION_TYPE_DESC)) {
+            validateInvalidReaonlyArguments(firstParamNode, secondParamNode);
+        } else if (firstParamSyntaxKind.equals(INTERSECTION_TYPE_DESC)) {
+            validateReadonlyWatchEventAndCaller(firstParamNode, secondParamNode, secondParamSyntaxKind);
+        } else if (secondParamSyntaxKind.equals(INTERSECTION_TYPE_DESC)) {
+            validateCallerAndReadonlyWatchEvent(firstParamNode, secondParamNode, firstParamSyntaxKind);
+        } else if (firstParamSyntaxKind.equals(QUALIFIED_NAME_REFERENCE)) {
+            validateInvalidArguments(firstParamNode, secondParamNode);
+        } else if (secondParamSyntaxKind.equals(QUALIFIED_NAME_REFERENCE)) {
+            validateInvalidArguments(secondParamNode, firstParamNode);
+        } else {
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_PARAMETERS,
                     DiagnosticSeverity.ERROR, functionDefinitionNode.functionSignature().location()));
         }
     }
 
-    private void validateIntersectionParam(ParameterNode parameterNode) {
+    private void validateSingleArgumentScenario(SeparatedNodeList<ParameterNode> parameters) {
+        ParameterNode paramNode = parameters.get(0);
+        SyntaxKind paramSyntaxKind = ((RequiredParameterNode) paramNode).typeName().kind();
+        if (paramSyntaxKind.equals(QUALIFIED_NAME_REFERENCE)) {
+            validateWatchEventParam(paramNode);
+        } else if (paramSyntaxKind.equals(INTERSECTION_TYPE_DESC)) {
+            boolean watchEventParam = validateIntersectionParam(paramNode);
+            if (!watchEventParam) {
+                context.reportDiagnostic(PluginUtils.getDiagnostic(
+                        INVALID_WATCHEVENT_PARAMETER,
+                        DiagnosticSeverity.ERROR, paramNode.location()));
+            }
+        } else {
+            context.reportDiagnostic(PluginUtils.getDiagnostic(
+                    INVALID_WATCHEVENT_PARAMETER,
+                    DiagnosticSeverity.ERROR, paramNode.location()));
+        }
+    }
+
+    private void validateInvalidArguments(ParameterNode firstParamNode, ParameterNode secondParamNode) {
+        boolean firstParamCaller = validateCallerParam(firstParamNode);
+        if (!firstParamCaller) {
+            validateWatchEventParam(firstParamNode);
+        }
+        context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_PARAMETERS,
+                DiagnosticSeverity.ERROR, secondParamNode.location()));
+    }
+
+    private void validateCallerAndReadonlyWatchEvent(ParameterNode firstParamNode, ParameterNode secondParamNode,
+                                                     SyntaxKind firstParamSyntaxKind) {
+        boolean firstParamCaller = validateCallerParam(firstParamNode);
+        boolean secondParamWatchEvent = validateIntersectionParam(secondParamNode);
+        if (!firstParamCaller && !secondParamWatchEvent) {
+            if (firstParamSyntaxKind.equals(QUALIFIED_NAME_REFERENCE)) {
+                validateWatchEventParam(firstParamNode);
+            }
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_PARAMETERS,
+                    DiagnosticSeverity.ERROR, secondParamNode.location()));
+        } else if (!secondParamWatchEvent) {
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_WATCHEVENT_PARAMETER,
+                    DiagnosticSeverity.ERROR, firstParamNode.location()));
+        } else if (!firstParamCaller) {
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_CALLER_PARAMETER,
+                    DiagnosticSeverity.ERROR, secondParamNode.location()));
+        }
+    }
+
+    private void validateReadonlyWatchEventAndCaller(ParameterNode firstParamNode, ParameterNode secondParamNode,
+                                                     SyntaxKind secondParamSyntaxKind) {
+        boolean firstParamWatchEvent = validateIntersectionParam(firstParamNode);
+        boolean secondParamCaller = validateCallerParam(secondParamNode);
+        if (!firstParamWatchEvent && !secondParamCaller) {
+            if (secondParamSyntaxKind.equals(QUALIFIED_NAME_REFERENCE)) {
+                validateWatchEventParam(secondParamNode);
+            }
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_PARAMETERS,
+                    DiagnosticSeverity.ERROR, firstParamNode.location()));
+        } else if (!firstParamWatchEvent) {
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_WATCHEVENT_PARAMETER,
+                    DiagnosticSeverity.ERROR, firstParamNode.location()));
+        } else if (!secondParamCaller) {
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_CALLER_PARAMETER,
+                    DiagnosticSeverity.ERROR, secondParamNode.location()));
+        }
+    }
+
+    private void validateInvalidReaonlyArguments(ParameterNode firstParamNode, ParameterNode secondParamNode) {
+        boolean firstParamWatchEvent = validateIntersectionParam(firstParamNode);
+        if (!firstParamWatchEvent) {
+            boolean secondParamWatchEvent = validateIntersectionParam(secondParamNode);
+            if (!secondParamWatchEvent) {
+                context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_WATCHEVENT_PARAMETER,
+                        DiagnosticSeverity.ERROR, secondParamNode.location()));
+            }
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_CALLER_PARAMETER,
+                    DiagnosticSeverity.ERROR, firstParamNode.location()));
+        } else {
+            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_CALLER_PARAMETER,
+                    DiagnosticSeverity.ERROR, secondParamNode.location()));
+        }
+    }
+
+    private void validateCallerAndWachEvent(ParameterNode firstParamNode, ParameterNode secondParamNode) {
+        boolean firstParamCaller = validateCallerParam(firstParamNode);
+        if (firstParamCaller) {
+            validateWatchEventParam(secondParamNode);
+        } else {
+            validateWatchEventParam(firstParamNode);
+            boolean secondParamCaller = validateCallerParam(secondParamNode);
+            if (!secondParamCaller) {
+                context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_CALLER_PARAMETER,
+                        DiagnosticSeverity.ERROR, secondParamNode.location()));
+            }
+        }
+    }
+
+    private boolean validateIntersectionParam(ParameterNode parameterNode) {
         RequiredParameterNode requiredParameterNode = (RequiredParameterNode) parameterNode;
         SemanticModel semanticModel = context.semanticModel();
         Optional<Symbol> symbol = semanticModel.symbol(requiredParameterNode);
@@ -116,18 +239,12 @@ public class FtpFunctionValidator {
                         .anyMatch(typeSymbol -> typeSymbol.nameEquals(PluginConstants.WATCHEVENT) &&
                                 typeSymbol.getModule().isPresent() &&
                                 validateModuleId(typeSymbol.getModule().get()));
-                if (!watchEventExists) {
-                    context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_PARAMETER,
-                            DiagnosticSeverity.ERROR, requiredParameterNode.location()));
+                if (watchEventExists) {
+                    return true;
                 }
-            } else {
-                context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_PARAMETER,
-                        DiagnosticSeverity.ERROR, requiredParameterNode.location()));
             }
-        } else {
-            context.reportDiagnostic(PluginUtils.getDiagnostic(INVALID_PARAMETER,
-                    DiagnosticSeverity.ERROR, requiredParameterNode.location()));
         }
+        return false;
     }
 
     private void validateWatchEventParam(ParameterNode parameterNode) {
@@ -143,19 +260,39 @@ public class FtpFunctionValidator {
                 if (!validateModuleId(moduleSymbol.get()) ||
                         !paramName.equals(PluginConstants.WATCHEVENT)) {
                     context.reportDiagnostic(PluginUtils.getDiagnostic(
-                            INVALID_PARAMETER,
+                            INVALID_WATCHEVENT_PARAMETER,
                             DiagnosticSeverity.ERROR, requiredParameterNode.location()));
                 }
             } else {
                 context.reportDiagnostic(PluginUtils.getDiagnostic(
-                        INVALID_PARAMETER,
+                        INVALID_WATCHEVENT_PARAMETER,
                         DiagnosticSeverity.ERROR, requiredParameterNode.location()));
             }
         } else {
             context.reportDiagnostic(PluginUtils.getDiagnostic(
-                    INVALID_PARAMETER,
+                    INVALID_WATCHEVENT_PARAMETER,
                     DiagnosticSeverity.ERROR, requiredParameterNode.location()));
         }
+    }
+
+    private boolean validateCallerParam(ParameterNode parameterNode) {
+        RequiredParameterNode requiredParameterNode = (RequiredParameterNode) parameterNode;
+        Node parameterTypeNode = requiredParameterNode.typeName();
+        SemanticModel semanticModel = context.semanticModel();
+        Optional<Symbol> paramSymbol = semanticModel.symbol(parameterTypeNode);
+        if (paramSymbol.isPresent()) {
+            Optional<ModuleSymbol> moduleSymbol = paramSymbol.get().getModule();
+            if (moduleSymbol.isPresent()) {
+                String paramName = paramSymbol.get().getName().isPresent() ?
+                        paramSymbol.get().getName().get() : "";
+                if (!validateModuleId(moduleSymbol.get()) ||
+                        !paramName.equals(PluginConstants.CALLER)) {
+                    return false;
+                }
+                return true;
+            }
+        }
+        return false;
     }
 
     private void validateReturnTypeErrorOrNil(FunctionDefinitionNode functionDefinitionNode) {
@@ -172,9 +309,9 @@ public class FtpFunctionValidator {
                                 if (!returnType.signature().equals(PluginConstants.ERROR) &&
                                         !validateModuleId(returnType.getModule().get())) {
                                     context.reportDiagnostic(PluginUtils.getDiagnostic(
-                                            PluginConstants.CompilationErrors.INVALID_RETURN_TYPE_ERROR_OR_NIL,
-                                            DiagnosticSeverity.ERROR,
-                                            functionDefinitionNode.functionSignature().location()));
+                                        PluginConstants.CompilationErrors.INVALID_RETURN_TYPE_ERROR_OR_NIL,
+                                        DiagnosticSeverity.ERROR,
+                                        functionDefinitionNode.functionSignature().location()));
                                 }
                             } else if (returnType.typeKind() != TypeDescKind.ERROR) {
                                 context.reportDiagnostic(PluginUtils.getDiagnostic(
@@ -185,8 +322,8 @@ public class FtpFunctionValidator {
                     }
                 } else if (returnTypeDesc.get().typeKind() != TypeDescKind.NIL) {
                     context.reportDiagnostic(PluginUtils.getDiagnostic(
-                            PluginConstants.CompilationErrors.INVALID_RETURN_TYPE_ERROR_OR_NIL,
-                            DiagnosticSeverity.ERROR, functionDefinitionNode.functionSignature().location()));
+                        PluginConstants.CompilationErrors.INVALID_RETURN_TYPE_ERROR_OR_NIL,
+                        DiagnosticSeverity.ERROR, functionDefinitionNode.functionSignature().location()));
                 }
             }
         }