+++ native/src/main/java/io/ballerina/stdlib/ftp/server/FtpListenerHelper.java
@@ -19,6 +19,8 @@
 package io.ballerina.stdlib.ftp.server;
 
 import io.ballerina.runtime.api.Runtime;
+import io.ballerina.runtime.api.creators.ValueCreator;
+import io.ballerina.runtime.api.types.MethodType;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BError;
 import io.ballerina.runtime.api.values.BMap;
@@ -31,12 +33,18 @@ import io.ballerina.stdlib.ftp.transport.impl.RemoteFileSystemConnectorFactoryIm
 import io.ballerina.stdlib.ftp.transport.server.connector.contract.RemoteFileSystemServerConnector;
 import io.ballerina.stdlib.ftp.util.FtpConstants;
 import io.ballerina.stdlib.ftp.util.FtpUtil;
+import io.ballerina.stdlib.ftp.util.ModuleUtils;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 
+import static io.ballerina.stdlib.ftp.util.FtpConstants.FTP_CALLER;
+import static io.ballerina.stdlib.ftp.util.FtpConstants.FTP_CLIENT;
+import static io.ballerina.stdlib.ftp.util.FtpConstants.FTP_SERVICE_ENDPOINT_CONFIG;
 import static io.ballerina.stdlib.ftp.util.FtpUtil.ErrorType.Error;
+import static io.ballerina.stdlib.ftp.util.FtpUtil.getOnFileChangeMethod;
 
 /**
  * Helper class for listener functions.
@@ -61,6 +69,8 @@ public class FtpListenerHelper {
                     .createServerConnector(paramMap, listener);
             ftpListener.addNativeData(FtpConstants.FTP_SERVER_CONNECTOR, serverConnector);
             // This is a temporary solution
+
+            ftpListener.addNativeData(FTP_SERVICE_ENDPOINT_CONFIG, serviceEndpointConfig);
             return null;
         } catch (RemoteFileSystemConnectorException | BallerinaFtpException e) {
             Throwable rootCause = findRootCause(e);
@@ -74,6 +84,23 @@ public class FtpListenerHelper {
                 FtpConstants.FTP_SERVER_CONNECTOR);
         FtpListener listener = ftpConnector.getFtpListener();
         listener.addService(service);
+
+        Optional<MethodType> methodType = getOnFileChangeMethod(service);
+        if (methodType.isEmpty() || methodType.get().getParameters().length != 2) {
+            return null;
+        }
+        if (listener.getCaller() != null) {
+            return null;
+        }
+        BMap serviceEndpointConfig  = (BMap) ftpListener.getNativeData(FTP_SERVICE_ENDPOINT_CONFIG);;
+        BObject caller = createCaller(serviceEndpointConfig);
+        if (caller instanceof BError) {
+            BError endpointError = (BError) caller;
+            return FtpUtil.createError(endpointError.getMessage(), endpointError.getDetails().toString(),
+                    Error.errorType());
+        } else {
+            listener.setCaller(caller);
+        }
         return null;
     }
 
@@ -154,4 +181,8 @@ public class FtpListenerHelper {
         return null;
     }
 
+    private static BObject createCaller(BMap<BString, Object> serviceEndpointConfig) {
+        BObject client = ValueCreator.createObjectValue(ModuleUtils.getModule(), FTP_CLIENT, serviceEndpointConfig);
+        return ValueCreator.createObjectValue(ModuleUtils.getModule(), FTP_CALLER, client);
+    }
 }