+++ native/src/main/java/io/ballerina/stdlib/ftp/util/FtpUtil.java
@@ -21,6 +21,7 @@ package io.ballerina.stdlib.ftp.util;
 import io.ballerina.runtime.api.Module;
 import io.ballerina.runtime.api.creators.ErrorCreator;
 import io.ballerina.runtime.api.creators.ValueCreator;
+import io.ballerina.runtime.api.types.MethodType;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BError;
@@ -40,11 +41,14 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Stream;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
 import static io.ballerina.stdlib.ftp.util.FtpConstants.FTP_ANONYMOUS_PASSWORD;
 import static io.ballerina.stdlib.ftp.util.FtpConstants.FTP_ANONYMOUS_USERNAME;
+import static io.ballerina.stdlib.ftp.util.FtpConstants.ON_FILE_CHANGE_REMOTE_FUNCTION;
 import static io.ballerina.stdlib.ftp.util.ModuleUtils.getModule;
 
 /**
@@ -208,6 +212,13 @@ public class FtpUtil {
 
     public static String getCompressedFileName(String fileName) {
         return fileName.substring(0, fileName.lastIndexOf('.')).concat(".zip");
+    }
+
+    public static Optional<MethodType> getOnFileChangeMethod(BObject service) {
+        MethodType[] methodTypes = service.getType().getMethods();
+        return Stream.of(methodTypes)
+                .filter(methodType -> ON_FILE_CHANGE_REMOTE_FUNCTION.equals(methodType.getName()))
+                .findFirst();
     }
 
     /**