+++ changelog.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
+ - [[#2525] Introduced `ftp:Caller` API and added as an optional parameter in `onFileChange` method](https://github.com/ballerina-platform/ballerina-standard-library/issues/2525)
  - [[#1514] Added compiler plugin validation support for ftp service](https://github.com/ballerina-platform/ballerina-standard-library/issues/1514)
  - [[#2704] Added code snippet generation for ftp service](https://github.com/ballerina-platform/ballerina-standard-library/issues/2704)